Trying again one more time to prevent import cycles.. Moved common.ExpandUser into packer.ExpandUser..
